+++ src/organisms/Header/HeaderContainer.js
@@ -1,25 +1,8 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import HeaderAppBar from "../molecules/HeaderAppBar";
-import { useAuth0 } from "../react-auth0-spa";
+import HeaderComponent from "./HeaderComponent";
+import { useAuth0 } from "../../react-auth0-spa";
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 0,
-  },
-  title: {
-    width: "20%",
-    color: "#ffffff",
-    border: "0",
-  },
-  button: {
-    marginRight: theme.spacing(2),
-    color: "#ffffff",
-  },
-}));
-
-export default function Header() {
-  const classes = useStyles();
+export default function HeaderContainer() {
   const { user, isAuthenticated, loginWithRedirect, logout } = useAuth0();
 
   // eslint-disable-next-line
@@ -29,13 +12,11 @@ export default function Header() {
     });
 
   return (
-    <div className={classes.root}>
-      <HeaderAppBar
-        user={user}
-        isAuthenticated={isAuthenticated}
-        loginWithRedirect={loginWithRedirect}
-        logoutWithRedirect={logoutWithRedirect}
-      />
-    </div>
+    <HeaderComponent
+      user={user}
+      isAuthenticated={isAuthenticated}
+      loginWithRedirect={loginWithRedirect}
+      logoutWithRedirect={logoutWithRedirect}
+    />
   );
 }
